Fix M12: Disturbing Presence now applies to Throw Team-mate PA tests
The BB2025 DP text explicitly names the Throw Team-mate Action, but
ttm_advance computed the PA test fully inline — no bb_ctx, so no
registered mod, aura or re-roll hook could ever touch a TTM throw.
The PA test now builds a BB_TEST_TTM ctx routed through bb_hook_mods
(also the future seam for TTM re-roll windows), and the DP aura filter
accepts TTM alongside PASS/CATCH. Accurate/Cannoneer and the Pass
re-roll stay excluded: they gate on BB_TEST_PASS and a TTM throw is
not a Pass Action.

Prerequisite hardening (review LOW note): bb_test_kind reached 9
values while reroll_kinds/skill_rr_used were uint8 masks — 1u << 8
truncated to 0. Both masks widened to uint16 with a
_Static_assert(BB_TEST_KIND_COUNT <= 16) so the new BB_TEST_TTM kind
(bit 9) cannot silently alias bit 0. (adversarial review M12)

+++ engine/include/bb/bb_hooks.h
@@ -54,7 +54,7 @@ typedef struct {
     bb_mod_fn armour_mod; // armour-roll modifier (c->player = downed player,
                           // c->other = causer; positive helps the causer)
     bb_mod_fn injury_mod; // injury-roll modifier (same convention)
-    uint8_t reroll_kinds; // bitmask of bb_test_kind this skill lets you re-roll
+    uint16_t reroll_kinds; // bitmask of bb_test_kind this skill lets you re-roll
     uint8_t activate_gate; // D6 target for the activation negatrait (0 = none)
     uint8_t gate_kind;     // bb_gate_kind behavior on failure
     uint8_t push_flags;    // BB_PUSHF_* (owner's effect in pushes)
